requirements.txt
Flask
sqlalchemy
pymysql
Flask-SQLAlchemy
pytest
python-dotenv
psycopg2-binary
psycopg2
